move home component to pages dir
