+++ client/src/pages/Home/gyb_logo.png
